requirements.txt
praw
groq
python-dotenv
